# docs/generated/api/swagger.yaml
definitions:
...
  handlers.AgentCreateRequest:
    properties:
      department:
        example: sales
        type: string
      max_concurrent:
        example: 5
        type: integer
      skills:
        example: tech,support
        type: string
      user_id:
        example: 1
        type: integer
    required:
    - user_id
    type: object
  handlers.AgentInfo:
    properties:
      avg_response_time:
        example: 120
        type: integer
      connected_at:
        example: "2024-01-01T10:00:00Z"
        type: string
      current_load:
        example: 2
        type: integer
      department:
        example: sales
        type: string
      last_activity:
        example: "2024-01-01T12:00:00Z"
        type: string
      max_concurrent:
        example: 5
        type: integer
      name:
        example: John Doe
        type: string
      rating:
        example: 4.8
        type: number
      skills:
        example:
        - tech
        - support
        items:
          type: string
        type: array
      status:
        example: online
        type: string
      user_id:
        example: 1
        type: integer
      username:
        example: john_doe
        type: string
    type: object
  handlers.AgentStats:
    properties:
      avg_rating:
        example: 4.5
        type: number
      avg_response_time:
        example: 120
        type: integer
      busy:
        example: 3
        type: integer
      online:
        example: 5
        type: integer
      total:
        example: 10
        type: integer
    type: object
  handlers.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
...
  models.Agent:
    properties:
      avg_response_time:
        description: 平均响应时间(秒)
        type: integer
      created_at:
        type: string
      current_load:
        description: 当前工单数
        type: integer
      department:
        type: string
      id:
        type: integer
      max_concurrent:
        description: 最大并发工单数
        type: integer
      rating:
        description: 评分
        type: number
      skills:
        description: 技能标签，逗号分隔
        type: string
      status:
        description: online, offline, busy
        type: string
      tenant_id:
        type: string
      tickets:
        items:
          $ref: '#/definitions/models.Ticket'
        type: array
      total_tickets:
        description: 总处理工单数
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
      workspace_id:
        type: string
    type: object
  models.CustomField:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      id:
        type: integer
      key:
        description: stable identifier (slug)
        type: string
      name:
        type: string
      options_json:
        description: JSON array (for select/multiselect)
        type: string
      required:
        type: boolean
      resource:
        description: ticket
        type: string
      show_when_json:
        description: JSON condition expression
        type: string
      tenant_id:
        type: string
      type:
        description: string, number, boolean, date, select, multiselect
        type: string
      updated_at:
        type: string
      validation_json:
        description: JSON object (min/max/regex/etc)
        type: string
      workspace_id:
        type: string
    type: object
...
  models.Message:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      sender:
        description: user, ai, agent
        type: string
      session:
        $ref: '#/definitions/models.Session'
      session_id:
        type: string
      tenant_id:
        type: string
      type:
        description: text, image, file, system
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
      workspace_id:
        type: string
    type: object
...
  models.Session:
    properties:
      agent:
        $ref: '#/definitions/models.User'
      agent_id:
        type: integer
      created_at:
        type: string
      ended_at:
        type: string
      id:
        type: string
      messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      platform:
        description: web, telegram, wechat, etc.
        type: string
      started_at:
        type: string
      status:
        description: active, ended, transferred
        type: string
      tenant_id:
        type: string
      ticket:
        $ref: '#/definitions/models.Ticket'
      ticket_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
      workspace_id:
        type: string
    type: object
  models.Ticket:
    properties:
      agent:
        $ref: '#/definitions/models.User'
      agent_id:
        type: integer
      attachments:
        items:
          $ref: '#/definitions/models.TicketFile'
        type: array
      category:
        description: technical, billing, general, complaint
        type: string
      closed_at:
        type: string
      comments:
        items:
          $ref: '#/definitions/models.TicketComment'
        type: array
      created_at:
        type: string
      custom_field_values:
        items:
          $ref: '#/definitions/models.TicketCustomFieldValue'
        type: array
      customer:
        allOf:
        - $ref: '#/definitions/models.User'
        description: 关联关系
      customer_id:
        type: integer
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      priority:
        description: low, normal, high, urgent
        type: string
      resolved_at:
        type: string
      session:
        $ref: '#/definitions/models.Session'
      session_id:
        type: string
      source:
        description: web, email, phone, chat
        type: string
      status:
        description: open, assigned, in_progress, resolved, closed
        type: string
      status_history:
        items:
          $ref: '#/definitions/models.TicketStatus'
        type: array
      tags:
        description: 标签，逗号分隔
        type: string
      tenant_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      workspace_id:
        type: string
    type: object
  models.TicketComment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      ticket:
        $ref: '#/definitions/models.Ticket'
      ticket_id:
        type: integer
      type:
        description: comment, internal_note, system
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.TicketCustomFieldValue:
    properties:
      created_at:
        type: string
      custom_field:
        $ref: '#/definitions/models.CustomField'
      custom_field_id:
        type: integer
      id:
        type: integer
      ticket_id:
        type: integer
      updated_at:
        type: string
      value:
        type: string
    type: object
  models.TicketFile:
    properties:
      created_at:
        type: string
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      id:
        type: integer
      mime_type:
        type: string
      ticket:
        $ref: '#/definitions/models.Ticket'
      ticket_id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.TicketStatus:
    properties:
      created_at:
        type: string
      from_status:
        type: string
      id:
        type: integer
      reason:
        type: string
      ticket:
        $ref: '#/definitions/models.Ticket'
      ticket_id:
        type: integer
      to_status:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
...
  models.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        description: customer, agent, admin
        type: string
      sessions:
        description: 关联关系
        items:
          $ref: '#/definitions/models.Session'
        type: array
      status:
        description: active, inactive, banned
        type: string
      tickets:
        items:
          $ref: '#/definitions/models.Ticket'
        type: array
      token_valid_after:
        type: string
      token_version:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/agents:
    get:
      consumes:
      - application/json
      description: 获取客服列表（用于批量指派下拉/后台管理）
      parameters:
      - description: 返回条数（默认 200，最大 500）
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Agent'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 获取客服列表
      tags:
      - 客服管理
    post:
      consumes:
      - application/json
      description: 创建新的客服代理
      parameters:
      - description: 客服信息
        in: body
        name: agent
        required: true
        schema:
          $ref: '#/definitions/handlers.AgentCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Agent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 创建客服
      tags:
      - 客服管理
  /api/agents/{id}:
    get:
      consumes:
      - application/json
      description: 根据用户ID获取客服的详细信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Agent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 获取客服详情
      tags:
      - 客服管理
  /api/agents/{id}/assign-session:
    post:
      consumes:
      - application/json
      description: 将会话分配给指定的客服
      parameters:
      - description: 客服用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 分配信息
        in: body
        name: assignment
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 分配会话给客服
      tags:
      - 客服管理
  /api/agents/{id}/offline:
    post:
      consumes:
      - application/json
      description: 将客服状态设置为离线
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 客服下线
      tags:
      - 客服管理
  /api/agents/{id}/online:
    post:
      consumes:
      - application/json
      description: 将客服状态设置为在线
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 客服上线
      tags:
      - 客服管理
  /api/agents/{id}/release-session:
    post:
      consumes:
      - application/json
      description: 从客服释放指定的会话
      parameters:
      - description: 客服用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 释放信息
        in: body
        name: release
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 释放客服的会话
      tags:
      - 客服管理
  /api/agents/{id}/revoke-tokens:
    post:
      consumes:
      - application/json
      description: 提升 token_version 并刷新 token_valid_after，使该客服旧 token 失效
      parameters:
      - description: 客服用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 强制失效客服 token
      tags:
      - 客服管理
  /api/agents/{id}/status:
    put:
      consumes:
      - application/json
      description: 更新客服的在线状态
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态信息
        in: body
        name: status
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 更新客服状态
      tags:
      - 客服管理
  /api/agents/find-available:
    get:
      consumes:
      - application/json
      description: 根据技能和优先级查找可用的客服
      parameters:
      - collectionFormat: csv
        description: 所需技能
        in: query
        items:
          type: string
        name: skills
        type: array
      - description: 优先级
        in: query
        name: priority
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AgentInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 查找可用客服
      tags:
      - 客服管理
  /api/agents/online:
    get:
      consumes:
      - application/json
      description: 获取当前在线的所有客服
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.AgentInfo'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 获取在线客服列表
      tags:
      - 客服管理
  /api/agents/stats:
    get:
      consumes:
      - application/json
      description: 获取客服相关的统计数据
      parameters:
      - description: 特定客服ID，用于获取单个客服的统计
        in: query
        name: agent_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AgentStats'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 获取客服统计
      tags:
      - 客服管理
...
swagger: "2.0"